make component beta with tests for dev
